+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-- If the player is sneaking (shift held), the Excavator will only break the single targeted block.
-- Added Argentine Spanish Language
+- Changed Stone Hammer Recipe